Implemented a simplifier for aperture macros. Already during the parse phase (actually when an aperture is defined using an aperture macro), gerbv now executes the byte code with the parameters given in the definition of the aperture. It generates a linked list with aperture macros with precalculated parameters, so the drawing engine never needs to redo the calculation everytime it draws an aperture defined by an aperture macro. And the virtual machine to execute the aperture macro byte code is only needed at one place (parser), not at every drawing engine.
The virtual machine is also made hardended against bugs in the byte
code.

An actual implementation of using these simplified aperture macro is also done
in the GDK drawing engine.

A bug in the GDK drawing engine is also fixed so when drawing aperture macro
primitive 20/2 so rotation is made in the correct way.