Fix Messages log tab missed messages
If file is opened not from gerbv GUI but form command line or from file
manager, then not all messages shown in Messages log tab. This log tab is not
ready, when such messages appears.

This patch put such messages to temporary buffer and output it when log tab is
ready.