Fix render bug for CW X-Y0 circles
The issue is that calculated floating point values should be zero, but are
actually very small with opposite signs. When put into atan2, that arc returns
alfa and beta of -pi and pi. The subsequent angle processing then manipulates
this into an incorrect, zero-length arc (pi to pi), instead of a (pi,3pi) arc
as would be expected.

This patch resolves the value by simply checking the values, and forcibly
declaring values below epsilon to be zero. The subsequent math then works
correctly, and the arc is handled correctly.

Impact of this bug:

This bug seems to only impact the rendering, and does not corrupt or mangle
files. So using GerbV with this bug will not result in production of a failed
board. It will, however, be very misleading, and potentially result in
confusion. Several design tools, notably Ultiboard, default to this arc
pattern for full circles, so it is a fairly common pattern to see in the wild.

Closes-bug: SF #68