Increase precision of exported RS274X files
Current PCB tools generally export gerbers in higher precision than GerbV
exports RS274X files in. In rare circumstances, the start and end coordinates
of an arc can be close enough that this loss of precision causes GerbV to round
them to the same coordinate. This makes the arclength 0, and the arc is then
treated as a full circle.

If this occurs near traces or ground planes, then it generally results in the
gerbers having a short, resulting in fabrication of the board to be incorrect.
Saving the 3.6 formatted gerbv with standard GerbV will result in the short
occurring.

This patch simply adjusts the export output to 3:6 format, which is equal to or
higher than most tools output. This avoids the precision loss, and thus the arc
rounding bug.

Closes-bug: SF #62