Use strcpy instead of strncpy to avoid warning.
The strncat is pointless because we already know the size so we can
just use strcpy.  strcat would work, too, but then the code would be
taking the length of the string once again and that is the [Schlemiel
Painter
algorithm](https://en.wikichip.org/wiki/schlemiel_the_painter%27s_algorithm).