Add support for Altium-style FILE_FORMAT comments.
Altium uses `FILE_FORMAT` comments to specify how many digits before
and after the decimal point.  The comment looks like this:

;FILE_FORMAT=4:4

This would indicate 4 digits before the decimal point and also after.
A prior or subequent `METRIC` or `INCH` command will specify whether
missing digits should be added to the left or the right by putting the
file into `LZ` mode or `TZ` mode.  In `LZ` mode, leading zeros are
required meaning that any number that doesn't have enough digits will
behave as if zeros are appended to the end until it has enough
digits.  In `TZ` mode, zeros are added to the front.

The last specified decimal format will be the one used.