fix: Turn off autod after FILE_FORMAT comment.
The `FILE_FORMAT` comment sets the number of digits before and after
the decimal.  If there is a following `INCH,TZ` command, we need to
disable the auto-detection so that this won't reset the digits to the
default, which is what usually happens when `INCH,TZ` isn't followed
by a format like `0000.0000`.