fix: Using floating point math for arcs.
Arcs and lines don't always line up when the image is zoomed out
because gdk expects integers so we do a lot of integer math.  For
example, under a lot of zoom out, we might be looking at arcs circles
with a diameter of 5.  When dividing by 2 to compute the radius, this
causes a problem.

The solution is to do as much floating-point math as possible and only
convert to integer at the very end, when we call the GDK functions to
draw.  In testing, this fixed #167.