build: Add a clang formatter.
This uses the `.clang-format` file to format any changed c or c++
code.

It uses whichever version of `clang-format` that can be found on the
compute that is newer than the specified major version.  The default
is to look for version 14 or later.  If a suitable `clang-format` is
not found, the program fails.

Only changes that are in the index (ie, that have been added by `git
add`) are formatted.  Changes in the worktree that aren't yet added to
the index are not formatted.

Any formatting changes that are found necessary are applied to both
the index and the current worktree.  This might fail if the index and
the worktree have diverged so significantly that the changes can't be
patched to the worktree.  This ought to be very rare.

A `-n` (equivalently, `-dry_run`) flag can be patched to the program
to cause the program to make no changes to any files.  Only the diff
will be printed out.  This diff can be applied by redirecting the
output of the dry-run to `patch -p1`.