`header_number_format` values restricted
Excellon format is clear that only two values
are valid for `header_number_format`, and
which value is used depends on the units:
* Inches ==> FMT_00_0000
* Metric ==> FMT_000_000

Therefore, do not set to FMT_USER on those
occasions, but to the allowed format type.

Also added many more comments to
`drill_parse_header_is_metric()`, explaining
the implied logic.