ci: Get ci working again.
We need to upgrade to ubuntu-latest.  ImageMagick is no longer
available by default so install that.  Also install bsdiff and
coreutils to be used by the unit test.

Due to the new compiler in ubuntu-latest, we need to make sure that
glib warnings will be ignored.  Also, there were some warnings about
tests in the code that were always true so those are removed.

The fonts used by the new Ubuntu don't match the old ones so we need
to update the golden files.  To make that easier, print the diff to
the log so that we can copy/paste it from the CI log.